requirements.txt
asgiref==3.8.1
Django==5.2
django-environ==0.12.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
